% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{bucket}
\alias{bucket}
\title{Retrieve Available Spatial Layers Metadata}
\usage{
bucket()
}
\value{
A data frame containing metadata fora ll available spatial layers
}
\description{
This function queries the Wildlife Move Tools API to obtain metadata for 
all available spatial layers in the Merkle Lab GIS system. The metadata 
includes details about each spatial layer, such as its name, description, 
type, and other relevant attributes.
}
\examples{
bucket()
}
