% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadDAYMET.R
\name{DownloadDAYMET}
\alias{DownloadDAYMET}
\title{Download DAYMET Daily Datasets From DEVISE Repository}
\usage{
DownloadDAYMET(params = c("prcp"), startDate, endDate, outDir)
}
\arguments{
\item{params}{A character vector containing a set of possible DAYMET datasets for download. prcp = Daily Precipitation in mm; swe = Snow Water Equivalent; tmax = Daily Maximum Temperature in Celsius; tmin = Daily Minimum Temperature in Celsius. Default is params = c("prcp", "swe", "tmax", "tmin")}

\item{startDate}{A character with the starting date of interest. ISO date format is required (e.g. YYYY-mm-dd)}

\item{endDate}{A character with the ending date of interest. ISO date format is required (e.g. YYYY-mm-dd). If NULL, only the single start date file will be downloaded.}

\item{outDir}{A character specifying the root output directory for downloaded files. This should be the root directory, sub-directories will be created if missing. Assumes your current working directory by default}
}
\value{
Returns the downloaded DEVISE study area DAYMET GeoTif files. Output is projected into epsg:5072 and cropped to DEVISE Study Extent
}
\description{
Downloads daily subsets of masked Snow Pack Temperature (Celcius) provided by the Snow Data Assimilation System (PRISM), see: \url{ftp://prism.nacse.org/daily/}.
}
