% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadRAP.R
\name{DownloadRAP}
\alias{DownloadRAP}
\title{RAP Download Function}
\usage{
DownloadRAP(
  RAPmetric = c("Biomass_AnnualForbsGrasses", "Cover_BareGround"),
  years = 2022,
  outDir
)
}
\arguments{
\item{RAPmetric}{character vector of RAP metrics to use (e.g. Biomass_AnnualForbsGrasses, Cover_Bareground)}

\item{years}{numeric vector of years for which you want to download the data}

\item{outDir}{character string representing the path to where you want to download the rasters}
}
\description{
The DownloadRAP function allows users to download specific metrics from the Rangeland Analysis Platform (RAP) for a given year.
}
