% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadSNODAS.R
\name{DownloadSNODAS}
\alias{DownloadSNODAS}
\title{Download SNODAS Daily Datasets From DEVISE Repository}
\usage{
DownloadSNODAS(params = c("swe", "snowdepth"), startDate, endDate, outDir)
}
\arguments{
\item{params}{A character vector containing a set of possible SNODAS datasets for download. SWE = Snow Water Equivalent and SnowDepth Default is params = c("SWE", "SnowDepth")}

\item{startDate}{A character with the starting date of interest. ISO date format is required (e.g. YYYY-mm-dd)}

\item{endDate}{A character with the ending date of interest. ISO date format is required (e.g. YYYY-mm-dd). If NULL, only the single start date file will be downloaded.}

\item{outDir}{A character specifying the root output directory for downloaded files. This should be the root directory, sub-directories will be created if missing. Assumes your current working directory by default}
}
\value{
Returns the downloaded DEVISE study area SNODAS GeoTif files. Output is projected into epsg:5072 and cropped to DEVISE Study Extent
}
\description{
Downloads daily subsets of masked raster data provided by the Snow Data Assimilation System (SNODAS), see: \url{https://nsidc.org/data/g02158}.
}
