% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractAnnualDAYMET.R
\name{ExtractAnnualDAYMET}
\alias{ExtractAnnualDAYMET}
\title{DAYMET Extraction Function}
\usage{
ExtractAnnualDAYMET(
  point_data,
  start_date,
  end_date,
  metric_name = c("prcp"),
  num_cores = NULL
)
}
\arguments{
\item{point_data}{sf point object}

\item{metric_name}{character vector of DAYMET metrics to use ("prcp", "swe", "tmax", "tmin")}
}
\value{
Returns a vector with length of point_data with the point/date values of DAYMET metrics
}
\description{
The Extract DAYMET function processes annual data from DAYMET
to extract specific metrics for given geographic points and dates.Users provide point spatial data
, desired DAYMET metrics ("Maxprcp","Maxswe","Maxtmax","Maxtmin","Meanprcp","Meanswe","Meantmax","Meantmin","Medianprcp","Medianswe","Mediantmax","Mediantmin"), and a date range. It extracts metric values for the provided points.
}
