% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractAnnualSNODAS.R
\name{ExtractAnnualSNODAS}
\alias{ExtractAnnualSNODAS}
\title{Annual SNODAS Extraction Function}
\usage{
ExtractAnnualSNODAS(
  point_data,
  start_date,
  end_date,
  metric_name = c("snowdepth"),
  num_cores = NULL
)
}
\arguments{
\item{point_data}{sf point object}

\item{metric_name}{character vector of SNODAS metrics to use (e.g. MaxSnowDepth, MeanSnowDepth)}
}
\value{
Returns a vector with length of XYdata with the point/date values of SNODASMetrics
}
\description{
The Extract Annual SNODAS function processes data from SNODAS
to extract specific metrics for given geographic points and dates.Users provide point spatial data
, desired SNODAS metrics (e.g., MaxSnowDepth, MaxSWE, MeanSnowDepth), and a date range. It extracts metric values for the provided points.
}
