% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractDailyDAYMET.R
\name{ExtractDailyDAYMET}
\alias{ExtractDailyDAYMET}
\title{Daily DAYMET Extraction Function}
\usage{
ExtractDailyDAYMET(
  XYdata,
  datesname = "date",
  Metrics = "prcp",
  num_cores = NULL
)
}
\arguments{
\item{XYdata}{sf point object}

\item{datesname}{name of your column representing date as POSIXct}

\item{Metrics}{character vector of DAYMET metrics to use ("prcp", "swe", "tmax", "tmin")}

\item{num_cores}{number of cores to use for parallel processing}
}
\value{
Returns a vector with length of XYdata with the point/date values of DaymetMetrics
}
\description{
The Extract Daily DAYMET function processes data from DAYMET
to extract specific metrics for given geographic points and dates.Users provide point spatial data
, desired DAYMET metrics either "prcp", "swe", "tmax", "tmin", and a date range. It extracts metric values for the provided points.
}
