% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractDailySNODAS.R
\name{ExtractDailySNODAS}
\alias{ExtractDailySNODAS}
\title{Daily SNODAS Extraction Function}
\usage{
ExtractDailySNODAS(
  XYdata,
  datesname = "date",
  Metrics = c("swe", "snowdepth"),
  num_cores = NULL
)
}
\arguments{
\item{XYdata}{sf point object}

\item{datesname}{name of your column representing date as POSIXct}

\item{Metrics}{character vector of SNODAS metrics to use (SWE or SnowDepth)}

\item{num_cores}{number of cores to use for parallel processing}
}
\value{
Returns a vector with length of XYdata with the point/date values of SNODASMetrics
}
\description{
The Extract Daily SNODAS function processes data from SNODAS
to extract specific metrics for given geographic points and dates.Users provide point spatial data
, desired SNODAS metrics either SWE or SnowDepth, and a date range. It extracts metric values for the provided points.
}
