% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_git2rdata.R
\name{ExtractRAP}
\alias{ExtractRAP}
\title{Extract RAP Metrics}
\usage{
ExtractRAP <- function(XYdata = d_rsf, 
                       RAPmetric = c("Biomass_AnnualForbsGrasses", "Cover_BareGround"),
                       bio_year_start = 150,
                       datesname = "date",
                       maxcpus = detectCores() - 1
}
\arguments{
\item{XYdata}{sf point object}

\item{RAPmetric}{Character vector of RAP metrics to use ( Biomass_AnnualForbsGrasses, Biomass_PerennialForbsGrasses, Cover_AnnualForbsGrasses, Cover_BareGround, Cover_Litter, Cover_PerennialForbsGrasses, Cover_Shrubs, Cover_Trees)}

\item{bio_year_start}{Julian date at which the biological year or growing season starts. In other words,before this julian day of a given year, the previous year's rap metrics will be extracted }

\item{datesname}{Name of your column representing date as POSIXct}
\item{maxcpus}{How many cores to use}

}

\description{
The ExtractRAP function processes data from the Rangeland Analysis Platform (RAP) to extract specific metrics for given geographic points and dates.Users provide spatial data (XYdata), desired RAP metrics (e.g., Biomass, Cover), a date column name, and a bio year, determining the growing season. For each year and metric, it fetches the associated RAP raster data set and extracts metric values for the provided points. }
\examples{
#Sample data
set.seed(42) 

# Create data frame
pts <- data.frame(
  long = runif(5, -111, -104),
  lat = runif(5, 41, 45),
  dates = as.POSIXct(sample(seq(as.Date('2010/01/01'), as.Date('2022/12/31'), by="day"), 5, replace = TRUE), tz = "UTC")
)
pts_sf <- st_as_sf(pts, coords = c("long", "lat"), crs = 4326, agr = "constant")
rap <- ExtractRAP(pts_sf, RAPmetric = "Cover_BareGround", datesname = "dates", )


}

\concept{internal}