% Generated by roxygen2: do not edit by hand
\name{ExtractSNODAS}
\alias{ExtractSNODAS}
\title{Extract SNODAS Metrics}
\usage{
ExtractSNODAS <- function(point_data, 
                                start_date = "2005-01-01", 
                                end_date = "2010-01-01", 
                                metric_name = "Maxprcp")
}
\arguments{
\item{point_data}{sf point object}

\item{metric_name}{Character vector of SNODAS metrics to use ("MaxSnowDepth","MaxSWE","MeanSnowDepth","MeanSWE"         "MedianSnowDepth,"MedianSWE" )}

\item{start_date}{Start date }

\item{end_date}{End date}

}

\description{
The Extract SNODAS function processes data from SNODAS
to extract specific metrics for given geographic points and dates.Users provide point spatial data
desired SNODAS metrics (e.g., MaxSnowDepth, MaxSWE, MeanSnowDepth), and a date range. It extracts metric values for the provided points.
}
\examples{
#Sample data
points <- data.frame(id = c(1, 2), x = c(-108.36312, -109.36312),  y = c(45.54731, 45.54731)) %>%
  sf::st_as_sf(coords = c("x", "y"), crs = 4326)
  
MaxSnowDepth <- ExtractSNODAS(points, start_date = "2005-01-01", 
                        end_date = "2010-01-01", metric_name = "MaxSnowDepth")
}

\concept{internal}