% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropRasters.R
\name{CropRasters}
\alias{CropRasters}
\title{Crop Rasters}
\usage{
CropRasters(cog_urls, polygon_sf, output_folder, writeData = FALSE)
}
\arguments{
\item{cog_urls}{A vector of raster URLs. These URLS should point directly to the geotiff files that you intend to crop from the Merkle Database}

\item{polygon_sf}{An sf polygon object}

\item{output_folder}{Output directory path of the cropped rasters}

\item{writeData}{If true, save the cropped rasters to the output folder. If FALSE (default), rasters are returned in memory only}
}
\description{
CropRasters is an R function to crop a set of cloud-optimized geotiffs using a polygon. After cropping, the rasters are saved to a specificed output 
directory where the directory structure reflects the orginal COG URLs.
}
