% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNDVI.R
\name{ExtractNDVI}
\alias{ExtractNDVI}
\title{NDVI Extraction Function}
\usage{
ExtractNDVI(XYdata, NDVImetric, datesname, maxcpus = 4)
}
\arguments{
\item{XYdata}{sf point object}

\item{NDVImetric}{character vector of NDVI metrics to use (e.g., MaxNDVIDay, MaxIRGday).}

\item{datesname}{Name of the column in XYdata representing date as POSIXct.}

\item{maxcpus}{Specifies how many cores to use for parallel processing.}
}
\value{
Returns a dataframe with rows corresponding to XYdata and the extracted values of NDVImetric for each point/date combination.
}
\description{
The ExtractNDVI function processes data from the MODIS NDVI datasets
to extract specific NDVI metrics for given geographic points and dates. Provide spatial data 
(XYdata), desired NDVI metrics (e.g., MaxNDVIDay, MaxIRGday), and a date column name. 
For each year and metric, it fetches the associated NDVI raster dataset 
and extracts metric values for the provided points.
}
\examples{
# To use this function, you would typically follow the pattern:
# NDVI <- ExtractNDVI(XYdata = mySpatialPoints, NDVImetric = c("MaxNDVIDay", "MaxIRGday"), datesname = "myDateColumn", maxcpus = 4)

}
