% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractRAP.R
\name{ExtractRAP}
\alias{ExtractRAP}
\title{RAP Extraction Function}
\usage{
ExtractRAP(
  XYdata,
  RAPmetric = c("Biomass_AnnualForbsGrasses", "Cover_BareGround"),
  bio_year_start = NULL,
  datesname,
  maxcpus = NULL
)
}
\arguments{
\item{XYdata}{sf point object}

\item{RAPmetric}{character vector of rap metrics to use (e.g. Biomass_AnnualForbsGrasses, Cover_Bareground)}

\item{bio_year_start}{julian date at which the biological year or growing season starts. In other words, 
before this julian day of a given year, the previous year's rap metrics will be extracted}

\item{datesname}{name of your column representing date as POSIXct}

\item{maxcpus}{how many cores}
}
\value{
Returns a vector with length of XYdata with the point/date values of RAPmetrics
}
\description{
The ExtractRAP function processes data from the Rangeland Analysis Platform (RAP)
to extract specific metrics for given geographic points and dates.Users provide spatial data
(XYdata), desired RAP metrics (e.g., Biomass, Cover), a date column name, and a bio year,
determining the growing season. For each year and metric, it fetches the associated RAP 
raster data set and extracts metric values for the provided points.
}
